macOS: Refactor create_macos_framework.py (#54557)
This is a refactoring with no semantic changes.

This refactors the macOS framework creation code to be more readable, and extracts it to `sky_utils.py`.

While I was pulling this out, also generalised the code to not hardcode `FlutterMacOS.framework` in case we one day manage to generate the iOS and macOS frameworks with the same name.

This is a reland of flutter/engine#54546 (reverted in flutter/engine#54549), the original was reverted in order to revert flutter/engine#54543 (reverted in flutter/engine#54550), which was reverted because it failed to preserve symlinks while zipping the macOS framework. That patch has been relanded with a fix in flutter/engine#54555. This patch has been rebased to tip-of-tree for attempt two.

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style